backend/requirements.txt
websocket-client==1.6.1
python-dotenv==1.0.0
websockets==11.0.3
aiohttp==3.8.5
aiohttp-cors==0.7.0 